Move todo loading and saving into the store
This is very very cool!
